make original tests and bdd test from class pass
